% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{GeomNormalViolin}
\alias{GeomNormalViolin}
\title{GeomNormalViolin}
\format{An object of class \code{GeomNormalViolin} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\description{
GeomNormalViolin
}
\keyword{internal}
