% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{StatNormalViolin}
\alias{StatNormalViolin}
\title{StatNormViolin}
\format{An object of class \code{StatNormalViolin} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 5.}
\description{
StatNormViolin
}
\keyword{internal}
