% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{StatNormViolin}
\alias{StatNormViolin}
\title{StatNormViolin}
\format{An object of class \code{StatNormViolin} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\description{
StatNormViolin
}
\keyword{internal}
